Move FishNet's add fish form to /views
