+++ families/track_and_trade/client/src/main.js
@@ -22,10 +22,11 @@ require('../styles/main.scss')
 
 const m = require('mithril')
 
-const FishForm = require('./fish_form')
 const api = require('./services/api')
 const transactions = require('./services/transactions')
 const navigation = require('./components/navigation')
+
+const FishForm = require('./views/add_fish_form')
 const Dashboard = require('./views/dashboard')
 const LoginForm = require('./views/login_form')
 const SignupForm = require('./views/signup_form')
